+++ js/data.js
@@ -1,5 +1,53 @@
-import {createImage, PHOTOS_COUNT} from './data-images.js';
+const DESCRIPTIONS = [
+  'Что-то на богатом',
+  'Мне этот мир абсолютно понятен',
+  'Позитивная энергия',
+  'Спокойно, Маша, я Дубровский',
+  'Природы мистическая красота',
+  'На радио волнах',
+  'Я знаю точно: невозможное возможно',
+  'Творческая жилка',
+  'Меме',
+  'Ноу дескрипшион',
+];
 
-const createArray = () => Array.from({length: PHOTOS_COUNT}, createImage);
+const NAMES = [
+  'Богдан',
+  'Мухтар',
+  'Стася',
+  'Мирон',
+  'Фроня',
+  'Виша',
+  'Карли',
+  'Олег',
+  'Ханс',
+  'Лулу',
+];
 
-export {createArray};
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const PHOTOS_COUNT = 25;
+
+const COMMENTS = {
+  MIN: 0,
+  MAX: 30
+};
+
+const LIKES = {
+  MIN: 15,
+  MAX: 200
+};
+
+const AVATARS_NUMBER = {
+  MIN: 1,
+  MAX: 6
+};
+
+export {DESCRIPTIONS, NAMES, MESSAGES, PHOTOS_COUNT, COMMENTS, LIKES, AVATARS_NUMBER};
